rework as niche scheduled task
