+++ src/Core/Framework/DependencyInjection/app.xml
@@ -710,5 +710,18 @@
 
             <tag name="console.command"/>
         </service>
+
+        <service id="Shopware\Core\Framework\App\ScheduledTask\SystemHeartbeatTask">
+            <tag name="shopware.scheduled.task"/>
+        </service>
+
+        <service id="Shopware\Core\Framework\App\ScheduledTask\SystemHeartbeatHandler">
+            <argument type="service" id="scheduled_task.repository"/>
+            <argument type="service" id="logger"/>
+            <argument type="service" id="event_dispatcher"/>
+
+            <tag name="messenger.message_handler" />
+        </service>
+
     </services>
 </container>
