+++ tests/unit/Core/Framework/App/Event/SystemHeartbeatEventTest.php
@@ -1,10 +1,10 @@
 <?php declare(strict_types=1);
 
-namespace Shopware\Tests\Unit\Core\Framework\SystemCheck\Event;
+namespace Shopware\Tests\Unit\Core\Framework\App\Event;
 
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\TestCase;
-use Shopware\Core\Framework\SystemCheck\Event\SystemHeartbeatEvent;
+use Shopware\Core\Framework\App\Event\SystemHeartbeatEvent;
 use Shopware\Core\Framework\Webhook\AclPrivilegeCollection;
 
 /**
